@@ -1,9 +1,7 @@
 import 'package:mocktail/mocktail.dart';
 import 'package:stream_chat/src/client/retry_policy.dart';
 import 'package:stream_chat/src/client/retry_queue.dart';
-import 'package:stream_chat/src/core/models/event.dart';
-import 'package:stream_chat/src/core/models/message.dart';
-import 'package:stream_chat/src/event_type.dart';
+import 'package:stream_chat/stream_chat.dart';
 import 'package:test/test.dart';
 
 import '../mocks.dart';
@@ -15,8 +13,9 @@ void main() {
 
   setUpAll(() {
     final retryPolicy = RetryPolicy(
-      shouldRetry: (_, attempt, __) => attempt < 5,
-      retryTimeout: (_, attempt, __) => Duration(seconds: attempt),
+      shouldRetry: (_, __, error) {
+        return error is StreamChatNetworkError && error.isRetriable;
+      },
     );
     when(() => channel.client.retryPolicy).thenReturn(retryPolicy);
 
@@ -48,21 +47,32 @@ void main() {
       verifyNever(() => logger.info(any()));
     });
 
+    test('should throw if message state is not failed', () {
+      final message = Message(
+        id: 'test-message-id',
+        text: 'Sample message test',
+        state: MessageState.sent,
+      );
+      expect(() => retryQueue.add([message]), throwsA(isA<AssertionError>()));
+    });
+
     test('should return if queue already contains the message', () {
       final message = Message(
         id: 'test-message-id',
         text: 'Sample message test',
+        state: MessageState.sendingFailed,
       );
       retryQueue.add([message]);
       expect(() => retryQueue.add([message]), returnsNormally);
       // Called only for the first message
-      verify(() => logger.info('Adding 1 messages')).called(1);
+      verify(() => logger.info('Adding 1 messages to the queue')).called(1);
     });
 
     test('`.add` should add failed request to the queue', () async {
       final message = Message(
         id: 'test-message-id',
         text: 'Sample message test',
+        state: MessageState.sendingFailed,
       );
       retryQueue.add([message]);
       expect(retryQueue.hasMessages, isTrue);
