@@ -400,8 +400,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
   bool get _hasQuotedMessage =>
       _effectiveController.message.quotedMessage != null;
 
-  bool get _isEditing =>
-      _effectiveController.message.status != MessageSendingStatus.sending;
+  bool get _isEditing => !_effectiveController.message.state.isInitial;
 
   BoxBorder? _draggingBorder;
 
@@ -587,7 +586,7 @@ class StreamMessageInputState extends State<StreamMessageInput>
               child: Column(
                 mainAxisSize: MainAxisSize.min,
                 children: [
-                  if (_hasQuotedMessage)
+                  if (_hasQuotedMessage && !_isEditing)
                     // Ensure this doesn't show on web & desktop
                     PlatformWidgetBuilder(
                       mobile: (context, child) => child,
@@ -1386,10 +1385,12 @@ class StreamMessageInputState extends State<StreamMessageInput>
       skipEnrichUrl: skipEnrichUrl,
     );
 
-    if (shouldKeepFocus) {
-      FocusScope.of(context).requestFocus(_effectiveFocusNode);
-    } else {
-      FocusScope.of(context).unfocus();
+    if (mounted) {
+      if (shouldKeepFocus) {
+        FocusScope.of(context).requestFocus(_effectiveFocusNode);
+      } else {
+        FocusScope.of(context).unfocus();
+      }
     }
   }
 
