@@ -30,7 +30,7 @@ extension MessageEntityX on MessageEntity {
         localDeletedAt: localDeletedAt,
         id: id,
         type: type,
-        status: status,
+        state: MessageState.fromJson(jsonDecode(state)),
         command: command,
         parentId: parentId,
         quotedMessageId: quotedMessageId,
@@ -70,7 +70,7 @@ extension MessageX on Message {
         reactionScores: reactionScores,
         reactionCounts: reactionCounts,
         mentionedUsers: mentionedUsers.map(jsonEncode).toList(),
-        status: status,
+        state: jsonEncode(state),
         remoteUpdatedAt: remoteUpdatedAt,
         localUpdatedAt: localUpdatedAt,
         extraData: extraData,
