@@ -30,7 +30,7 @@ extension PinnedMessageEntityX on PinnedMessageEntity {
         localDeletedAt: localDeletedAt,
         id: id,
         type: type,
-        status: status,
+        state: MessageState.fromJson(jsonDecode(state)),
         command: command,
         parentId: parentId,
         quotedMessageId: quotedMessageId,
@@ -71,7 +71,7 @@ extension PMessageX on Message {
         reactionScores: reactionScores,
         reactionCounts: reactionCounts,
         mentionedUsers: mentionedUsers.map(jsonEncode).toList(),
-        status: status,
+        state: jsonEncode(state),
         remoteUpdatedAt: remoteUpdatedAt,
         localUpdatedAt: localUpdatedAt,
         extraData: extraData,
